requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Bcrypt
python-dotenv